% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbm.R
\name{lbm.control}
\alias{lbm.control}
\title{Auxiliary control for \code{lbm}}
\usage{
lbm.control(epsilon = 1e-8, maxit = 100)
}
\arguments{
\item{epsilon}{positive convergence tolerance \code{epsilon};}

\item{maxit}{integer giving the maximal number of iterations.}
}
\value{
A list with components named as the arguments.
}
\description{
Auxiliary function for \code{lbm} fitting. Only used
internally by \code{lbm}.
}
\seealso{
\code{glm.control}
}
